+++ tubular/base.py
@@ -4,6 +4,7 @@ from. These transformers contain key checks to be applied in all cases.
 """
 
 import pandas as pd
+import warnings
 from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.utils.validation import check_is_fitted
 
@@ -54,7 +55,7 @@ class BaseTransformer(TransformerMixin, BaseEstimator):
 
         if not isinstance(verbose, bool):
 
-            raise ValueError("verbose must be a bool")
+            raise TypeError("verbose must be a bool")
 
         else:
 
@@ -85,21 +86,21 @@ class BaseTransformer(TransformerMixin, BaseEstimator):
 
                     if not isinstance(c, str):
 
-                        raise ValueError(
-                            "each element of columns should be a single column name"
+                        raise TypeError(
+                            "each element of columns should be a single (string) column name"
                         )
 
                 self.columns = columns
 
             else:
 
-                raise ValueError(
+                raise TypeError(
                     "columns must be a string or list with the columns to be pre-processed (if specified)"
                 )
 
         if not isinstance(copy, bool):
 
-            raise ValueError("copy must be a bool")
+            raise TypeError("copy must be a bool")
 
         else:
 
@@ -134,27 +135,57 @@ class BaseTransformer(TransformerMixin, BaseEstimator):
 
         if y is not None:
 
-            if isinstance(y, pd.DataFrame):
+            if not isinstance(y, pd.Series):
 
-                if y.shape[1] > 1:
-
-                    raise ValueError(
-                        "multi-column DataFrame passed for y, expecting 1 column"
-                    )
-
-            elif isinstance(y, pd.Series):
-
-                pass
-
-            else:
-
-                raise ValueError("unexpected type for y")
+                raise TypeError("unexpected type for y, should be a pd.Series")
 
             if not y.shape[0] > 0:
 
                 raise ValueError(f"y is empty; {y.shape}")
 
         return self
+
+    def _combine_X_y(self, X, y):
+        """Combine X and y by adding a new column with the values of y to a copy of X.
+
+        The new column response column will be called `_temporary_response`.
+
+        This method can be used by transformers that need to use the response, y, together
+        with the explanatory variables, X, in their `fit` methods.
+
+        Parameters
+        ----------
+        X : pd.DataFrame
+            Data containing explanatory variables.
+
+        y : pd.Series
+            Response variable.
+
+        """
+
+        if not isinstance(X, pd.DataFrame):
+
+            raise TypeError("X should be a pd.DataFrame")
+
+        if not isinstance(y, pd.Series):
+
+            raise TypeError("y should be a pd.Series")
+
+        if X.shape[0] != y.shape[0]:
+
+            raise ValueError(
+                f"X and y have different numbers of rows ({X.shape[0]} vs {y.shape[0]})"
+            )
+
+        if not (X.index == y.index).all():
+
+            warnings.warn("X and y do not have equal indexes")
+
+        X_y = X.copy()
+
+        X_y["_temporary_response"] = y.values
+
+        return X_y
 
     def transform(self, X):
         """Base transformer transform method; checks X type (pandas DataFrame only) and copies data if requested.
@@ -216,7 +247,7 @@ class BaseTransformer(TransformerMixin, BaseEstimator):
 
         if not isinstance(X, pd.DataFrame):
 
-            raise ValueError("X should be a pd.DataFrame")
+            raise TypeError("X should be a pd.DataFrame")
 
         if self.columns is None:
 
@@ -224,7 +255,7 @@ class BaseTransformer(TransformerMixin, BaseEstimator):
 
         if not isinstance(self.columns, list):
 
-            raise ValueError("self.columns should be a list")
+            raise TypeError("self.columns should be a list")
 
         for c in self.columns:
 
@@ -246,7 +277,7 @@ class BaseTransformer(TransformerMixin, BaseEstimator):
 
         if not isinstance(X, pd.DataFrame):
 
-            raise ValueError("X should be a pd.DataFrame")
+            raise TypeError("X should be a pd.DataFrame")
 
         if self.columns is None: